requirements.txt
Django==4.2.16
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
